Analysis: Suppress same parameter value warning on reader comment list
Warning: "Actual value of parameter 'refreshing' is always 'false'"

------------------------------------------------------------------------

Currently 'setRefreshing(false)' is used three times and 'setRefreshing
(true)' is never used, but this doesn't mean it will not be used in the
future, or wasn't used in the past. As such, keeping this method
signature and suppressing this warning for the time being seems to be
the better approach here.