Refactor: Remove test only comment store method
The 'instantiateCommentModel(...)' related 'CommentStore' methods is
being in production and are only being used on a couple of connected
tests. Thus, it is better to remove those and not have them exposed and
available for clients like JP/WPAndroid to exploit.

PS: This change was done to assist with a future refactor and in order
to avoid extra testing.