Update Follow button status for recommended blog Discover cards
Observe the `followStatusUpdate` LiveData to instantly change the state of the
follow button for the recommended blogs card in the Discover feed. It also sets
the button as enabled or disabled based on the status updated.

Currently there's a small issue that makes the button be enabled even before the
final change in state for `followStatusUpdate` because the Discover feed is also
updated by another data source (`readerDiscoverDataProvider.discoverFeed`) which
is directly hooked to the `ReaderPostTableActionEnded` that is emitted by the
Event Bus when there are changes in the underlying table.

This issue is fairly small since the worst thing that can happen is the ripple
animation showing up but there are other safeguard in place to avoid making
duplicate request so the overall state does not get insconsistent.