Build: Update fluxc version to pr hash (#2870)
This 'FluxC' PR hash updates the library to that branch version
where the 'CommentModel' is updated to its new null proof version.

FluxC PR: https://github.com/wordpress-mobile/
WordPress-FluxC-Android/pull/2870

This step is required in order to check that these 'FluxC'
related changes work as expected for WPAndroid.

------------------------------------------------------------------------

FYI: This change is an addition to
ab1afaa82ef543195f05599fdf532e4427f9188b. This change is necessary in
order to overcome and resolve new merge conflicts.