Analysis: Suppress same parameter value warning on media store
Warning: "Actual value of parameter 'errorType' is always
'MediaErrorType.MALFORMED_MEDIA_ARG' "

------------------------------------------------------------------------

'notifyMediaUploadError(MediaErrorType.MALFORMED_MEDIA_ARG, ...)' is
currently used once and no other usage of this private method is
recorder, but this doesn't mean it will not be used in the
future, or wasn't used in the past. As such, keeping this method
signature and suppressing this warning for the time being seems to be
the better approach here.