Analysis: Add missing nn-a to multi field on media wp com rest response
FYI: 'nn-a' stands for 'non-null annotations'.

PS: This 'NotNullFieldNotInitialized' warning got suppressed because a
'response' class can never have its fields initialized via a constructor
initialization, or otherwise for that matter.

------------------------------------------------------------------------

Also, note that as part of this change, the 'response' that was
previously considered '@Nullable' has been updated to '@NonNull' as it
doesn't seem that a nullable response to be valid, or part of a specific
valid flow. The fact that 'MediaWPRESTResponse.kt' and
'BaseWPV2MediaRestClient.kt' is also working with a '@NonNull' response
give this change some extra confidence.