Show loading indicator even when there's posts on the feed
Right now, when the feed is not empty and we are loading posts, we don't show a
loading progress indicator and just silently fetch data. The issue is that when
fetch is finished the screen suddenly updates with the new posts and that looks
bad to the user as it didn't look like anything was going on.

This fixes it by always showing the progress indicator (swipe-to-refresh) even
when there are posts being shown.