Analysis: Suppress spread operator on page list event listener
Detekt Warning: "SpreadOperator: In most cases using a spread
operator causes a full copy of the array to be created before calling a
method. This may result in a performance penalty."

Now that this warning is suppressed close to source, it is no longer
needed to be in the baseline.