Build: Update fluxc version to pr hash (#2893)
This 'FluxC' PR hash updates the library to that branch version
where the 'Theme' payload classes are updated to their new null proof
version.

FluxC PR: https://github.com/wordpress-mobile/
WordPress-FluxC-Android/pull/2893

This step is required in order to check that these 'FluxC'
related changes work as expected for WPAndroid.

------------------------------------------------------------------------

FYI: This change is an addition to
6abaf1354e69b9428475ddfcb4bea38f33af42b9. This change is necessary in
order to overcome and resolve new merge conflicts.