Add check to ensure quality value is valid
Addresses bug reported here: https://github.com/wordpress-mobile/WordPress-Android/pull/19644/files#r1402179477

As we've made changes to the values in the quality array, there may be times when users have previously chosen values that no longer exist. Those cases were not being properly handled in the code.

With this commit, we now check if a quality value exists in the array. If it doesn't, we return the default quality value.