Rename and reposition tracks event for clarity
"app_settings_optimize_images_popup_changed" has been renamed to "app_settings_optimize_images_popup_tapped" to clarify the action the user's taken. Tapping the popup may not necessarily change the optimization value. This also brings the event in alignment with the naming used on iOS.

The list position of the tracks event has also been changed, so that it's grouped closer to other settings related to image optimization.