Hoist selected item state from JetpackDropdownMenu
`MenuElementData` needed to be refactored as well to support hoisting the state
for slected item, since having the click callback inside the item itself meant
it's not possible to instantiate a Single item AND set a callback function in it
that passes itself to some function call.

The change consisted of removing the click callback and adding an `id` to the
`Items` so they can be easily identifiable in callbacks passing it.