fix: Prevent Gutenberg connection logic from crashing Aztec
A network connectivity subscriber was added to the post activity in: https://github.com/wordpress-mobile/WordPress-Android/pull/19692

However, it was not scoped to only run in the Gutenberg editor. Because
no guard was in place, the subscriber attempted to invoke a non-existent
method on the Aztec editor, resulting in a crash.