Fix using null topBarUiState and add wait for a non-null value
The added code tries to wait until the topBarUiState is not null by polling for
its data for a few times. If it's still null it simply doesn't run the code, to
avoid calling topbar updates that should only happen if there is already a top
bar visible. If it's not null after that check (normal scenario) then it calls
the block and executes the logic.