Fix tag list check logic to avoid saving same data to DB
IMPORTANT: this breaks ReaderPostListFragment/ViewModel fetch logic because it
was relying on this bug causing FollowedTagsChanged to be triggered in the
EventBus all the time. With this fix the Subscriptions feed is completely broken
(and the other feeds that use this Fragments are likely to be broken as well,
which are all the other except Discover).