Fix empty view being hidden when coming back to Subscriptions
The empty view is being reused for showing several messages, including the fetch
message. Because of that, that are pieces of code that rely on previous code
showing the "empty view" and just alter its content, without calling "show"
again. If we hide the empty view when the list is actually empty, we get to
a scenario of a blank screen when coming back to this feed from some other
screens. Example: with 0 blog subscriptions, in the Subscriptions feed, going to
the tag suggestions and back cause the Subscriptions feed to be blank.