Only pass a single category to the Filters Sheet
The current implementation only ever expects to receive a single category but
to keep most of the old code we are still using a List of categories for the
sub filter logic. That is not needed on the roo SubfilterBottomSheetFragment
though and the presence of the list was causing a crash since the Parcelable
implementation of the SubfilterCategory was not working as expected.

Removing the list from that Fragment and using a single category lets us remove
the `Parcelable` requirement from the SubfilterCategory enum since we can now
deal with a simple Serializable class, therefore fixing the crash.