Fixing profile fields list constraint in MyProfileFragment
The list height was wrap_content, so when the space was limited with the new Gravatar view, the list was unusable. It kept behind the Gravatar view.

This was visible, for example, in landscape or with a bigger text size.

Adding the constraint to the top of the Gravatar view should minimize the issue.

There are still some corners case where the UX can be improved.