Remove duplicates from the server blogs list if local and remote doesn't match.
Remove duplicates from the server blogs list if local and remote doesn't match.
This is required because under obscure circumstances the server can return duplicates.
We could have modified the function isSameList to eliminate the length check, but it's better to keep it separate since we aim to remove this check as soon as possible.