Move Gravatar info banner from "MyProfileFragment" to "MeFragment"
After reviewing with designers, we want to show the informative banner on the first screen as users can modify their avatar in the "Profile" fragment.

The sync information banner will be shown in both fragments, MyProfile and MeFragment, after any modification on the Gravatar profile. The sync banner is reused, so we've extracted it to a separate XML layout.

This commit mainly moves the logic and view from the MeFragment to the MyProfile one.