Rename FollowedTagsChanged to FollowedTagsFetched
Upon investigation I realized that most places using this event were actually
using it as a callback for the end of the fetch task rather than something that
is ONLY emitted when the tags CHANGE, which was what the name implied.

That only worked because of the bug mentioned in commit 3017bab that did not
properly compare the local and server tags and considered all fetches as
requiring a local database write, which in turn triggered the event.

The constant triggering of that FollowedTagsChanged event was also the cause of
bugs #20009 and #20010 since part of the code that runs there was actually only
supposed to run some times.

To fix that I renamed it to FollowedTagsFetched to better represent what it does
and figured out which parts of the code inside `ReaderPostListFragment` event
listener were supposed to run and in which scenario. I also added a `didChange`
field in the `FollowedTagsFetched` event for completeness but that part might
not be needed in the future, since the only code using that is legacy code
related to the old subfilter which should be removed in the future.