Track the number of followed tags when the "FollowedTagsChanged" event is fired
Note that we're aware of an underlying issue with the FollowedTagsChanged event being fired excessively, even when the tags have not changed. To address this, I've implemented a guard to prevent excessive analytics tracking, currently set to 1 hour. This guard does not verify the previous number of tags sent to analytics. Considering the frequent triggering of "tags changed," this seems to be a reasonable compromise.