Remove work-arounds for buggy framework implementations
It seems that these issues are most likely related with issues in the
implementation of FragmentStatePagerAdapter, which is no longer used in
this Activity after migrating to ViewPager2. At least one improvement
can be found in FragmentStateAdapter, which is the use of
LongSparseArray, which possibly alleviates the pressure on the
state-saving mechanism.

Still, it would be good to keep an eye on the release that these changes
ship in to ensure it doens't introduce regressions for edge cases.