Fix toolbar icon color to match theme text color
The icon colors was being set manually via code because of collapsed vs expanded
cases, but it was using the Fragment's context, which resolves to the Activity's
context instead of the context I'm passing to the LayoutInflater (which is used
for all layout inflations, but doesn't override the Fragment/Activity context).

I was able to change the code to use `view.context` instead of just `context`
and also hardcoded the expanded color (when there's an image being shown) to
white, which was being done anyway but indirectly since we only had light and
dark modes.