Add test for ThemeValues (Ignored by default)
Unfortunately the ThemeValues class uses the `Color` Android class, which is not
available in the JVM during tests and all its methods just return 0, making the
test fail.

I thought about just mocking the implementation, but then a good part of this
test would be just testing mocks, which is not great. The "better" approach
would be running with a Robolectric Test Runner, but it looks like Robolectric
was removed from our project?