Change SubFilterViewModelOwner to ...Provider
The idea is that the Provider is the component responsible for fetching the
correct SubFilterViewModel, but it doesn't necessarily own the ViewModel.

This was done so we can use the current feed as the Owner of the SubFilter
ViewModel without it actually declaring it, so the outermost layer of the Reader
(ReaderFragment) is still the source of truth for getting the SubFilterViewModel
but the actual lifecycle of the ViewModel is tied to the current feed fragment.