Enable activity's auto instrumentation
For now, let's enable Activity's auto instrumentation for performance monitoring, as at this moment we don't have any other transactions defined. If we'll come to a point of having to many transactions, we can disable this feature again (as it's not optimal way of measuring performance transactions). Related discussion: https://github.com/wordpress-mobile/WordPress-Android/pull/18507\#issue-1724560137