Migrate SignupEpilogueFragment.java to Kotlin
Gravatar 1.0.0 contains suspend functions for uploading a new avatar to the Gravatar account. We allow users to upload avatars when creating their accounts in SignupEpilogueFragment.java, so we are using that method.

Using suspension functions from Java is far from ideal, so migrating the fragment to Kotlin will simplify the bump of Gravatar SDK.