Build: Ignored variants configuration for dependency analysis plugin
This commit updates the encrypted 'gradle.properties.enc' file to point
to that pinned hash which has the
'dependency.analysis.android.ignored.variants'
property added into the 'gradle.properties' file for the WPAndroid
project.

FYI: With this configuration only the 'debug' variant (build type) for
the lib modules and 'wordpressJalapenoDebug' + 'jetpackJalapenoDebug'
variants (product flavor + build type) for the app module is going to
be included, with the rest of the variants ignored.