ci: Fix Checkstyle configuration
Update configuration to address the following warning:

```
The Checkstyle rules file could not be parsed. cannot initialize module TreeWalker - TreeWalker is not allowed as a parent of LineLength Please review 'Parent Module' section for this Check in web documentation if Check is standard. The file has been blocked for 60s.
```