Resolve `violation-comments` plugin in main project build script
Because it's now resolved in the main build script, and not in the build script of `WordPress` module, it'll use dependencies provided by all other plugins that are resolved in the main build script as well.

So, in this particular case, `gson` in version `2.10.1` will be used for violations plugin, because `androidx.databinding:databinding-compiler-common:8.5.1` uses gson `2.10.1` and Gradle always selects the highest version.

Previously, it was not preferring `2.10.1`, because violations plugin was only resolved in `WordPress` module, and its build script was resolved in isolation from the main build script.