Deps: Remove androidx media dependency
This dependency is completely unused.

FYI: It got introduced as part of this
(82f8eaf2693e7b28b47b8adcfd9e7d83eeaf0e3f and then
252d89f522875b911c5eb97b1c1a68bc3d2f4423 and) commits, first an update
to the support libs to a newer version and then as part of the AndroidX
migration (not sure why).

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :WordPressLoginFlow
Unused dependencies which should be removed:
  implementation 'androidx.media:media:1.2.1'
  ...