Build: Replace kapt with annotation processor for fluxc processor
This is done in order to make this pure Java module ends-up becoming as
simple as possible and avoid tools like Dependency Analysis falsely
suggesting that there exist unused dependencies where there is none.

FYI: KAPT was added to this module via this
(6051d8437f6b13b3ba0b9fea09a5abb2ca519436) commit, because of the fact
that the 'AutoService' annotation processor got split into two packages.
The assumption here is that they could have used 'annotationProcessor'
back then, that is, instead of 'kapt' and adding an unnecessary
dependency to the 'kotlin' plugin altogether.