Add missing `environment` parameter to Buildkite trigger call
As noticed by @oguzkocer during the 25.6 release train.

Looking at the Git log, the `environment` hash was introduced in
b80c544e2756ce73b3c2f2c71a9ebe4134320df7 as part of adding support for a
different build trigger workflow when running in CI.

At the time, we did not notice that the pre-existing method to
trigger builds did not include environment variables. I guess the need
for env vars is something we didn't have before those changes, hence why
we never noticed the lack of it till now.

Notice that the iOS counterpart includes the `environment` dictionary,
https://github.com/wordpress-mobile/WordPress-iOS/blob/afd3a6533b61c71ec661d2621e9a3486aaadef01/fastlane/lanes/release.rb#L476-L500