Tests: Replace com.zendesk.sdk with zendesk.support import to resolve r
Updating to Zendesk 5.3.0 comes with singificant changes in the Zendesk
SDK structure and naming conventions. This cause an issue with the
'com.zendesk.sdk.R' import that was no longer resolvable. Updating the
import to 'zendesk.support.R' solves this issue.

For more info see:
- 5.3.0:
  https://developer.zendesk.com/documentation/classic-web-widget-sdks/
support-sdk/android/release_notes/#530
- Migration from Classic SDKs to Zendesk SDK:
  https://developer.zendesk.com/documentation/zendesk-web-widget-sdks/
sdks/android/migration/#migration-from-classic-sdks-to-zendesk-sdk