fix: Address editor crashes from null listeners (#21658)
* fix: Address editor crashes from null listeners

Setting the listeners within `instantiateItem` appears problematic as
this only runs for the initial launch, but not for other events that
destroy and recreate fragment--e.g., screen rotation events.

To address this, GutenbergKit now follow the pattern of the Aztec and
GutenbergMobile editors, invoking the `initializeEditorFragment` method.
This method is described as "requesting dependency injection," which
feels like an odd statement. There is likely further opportunity to
refactor this architecture for simplicity.

* docs: Add release note

* docs: Remove `TODO` comment

Apparently the linter prohibits this statement...