Deps: Stop ignoring the agp update
The AGP update no longer needs to be synced with other in-house
library and client projects (except for GBKit). As such, there is no
need for Dependabot to ignore them anymore.

This limitation got lifted due to the fact that:
- 'FluxC' got vendored into this project, and thus the need to sync the
AGP version with WCAndroid is no longer relevant (paaHJt-7hU-p2).
- It has been decided that any project that requires compatibility with
composite build on a library level, they will have to make the change
themselves, when and if needed (paaHJt-6Qv-p2).

------------------------------------------------------------------------

About the 'GBKit' exception, following the below PR's approach, a
product engineer should be able to replicate a future AGP update without
too much trouble:

PS: [Build] Update Gradle to 8.12.1 and AGP to 8.8.1 #21680
- https://github.com/wordpress-mobile/WordPress-Android/pull/21680

Thus, it is better to unblock such automated AGP updates, based on
Dependabot, instead of leaving them become stale for too long.