Workaround for lint crash
The link is crashing (not showing an error) because of some libraries incompatibility. The problem is that updating them is not trivial, as it involves upgrading the AGP as well. So, let's go with this workaround.
After internal conversation, next step is to change a bit the logic. So, hopefully flow will be used in a different way and lint wont crash.