fix: Mitigate GutenbergKit text composition oddities (#21883)
* fix: Mitigate Android text composition oddities

Forcibly ending text composition led to odd outcomes--e.g., erroneously
dismissing the inline inserter, resetting editor history stack.

It was originally added to ensure the latest text in the editor is
captured before providing it to the host app for persistence. This
appears necessary for Samsung device keyboards only. For the time being,
text may be lost on these devices if the user does not first blur the
text input before saving or exiting the editor.

* fix: End GutenbergKit text composition during close

Ensure the latest text is captured before attempting to persist the post
content.

* refactor: Improve error handling consistency

The rest of the editor methods throw this exception.

* build: Update GutenbergKit version

* style: Address lint warnings

* style: Address missing null annotations

* refactor: Reuse existing `doFinish` argument

The existing `doFinish` argument is used to managed finishing the
activity after saving. Because we want to complete text composition for
this same context, reusing this argument seems appropriate.

* build: Update GutenbergKit version