fix: Avoid unnecessary GutenbergKit starts
The editor relies upon a subscription to editor settings within the
FluxC store. FluxC can dispatch change events multiple times, leading to
unexpected, unnecessary invocations of GutenbergKit's `start` method.
This lead to odd outcomes.

First, during editor setup, we dispatch an event to request the latest
editor settings. FluxC broadcasts two change events: first with the
cached settings, and second when the fetched settings resolve. This
caused two `start` invocations when opening the editor.

Second, the My Site fragment requests the latest settings as a
performance optimization. When closing the editor and returning to My
Site, the request resulted in a broadcast of updated settings, which
attempt to start the editor while it was closing.