feat: Enable GutenbergKit asset caching (#22035)
* refactor: Remove redundant private Atomic site check

Using `isWPComAtomic` suffices for the need of identifying sites managed
by WPCOM.

* feat: Cache GutenbergKit assets and improve configuration

* fix: Disable WebView preloading causing GutenbergKit load failures

The current preloading strategy results in the editor loading twice. The
duplicative load results in hidden editor UI for the second load, as the
loading sequence is not idempotent currently. We should refactor to
ensure the editor UI is only displayed once after the editor fully
loads.

Currently, The first occurs without blog configuration and always loads
the bundled editor. We should ensure preload loads with the proper
configuration and the correct editor (bundled/remote).

* fix: Avoid unnecessary GutenbergKit starts

The editor relies upon a subscription to editor settings within the
FluxC store. FluxC can dispatch change events multiple times, leading to
unexpected, unnecessary invocations of GutenbergKit's `start` method.
This lead to odd outcomes.

First, during editor setup, we dispatch an event to request the latest
editor settings. FluxC broadcasts two change events: first with the
cached settings, and second when the fetched settings resolve. This
caused two `start` invocations when opening the editor.

Second, the My Site fragment requests the latest settings as a
performance optimization. When closing the editor and returning to My
Site, the request resulted in a broadcast of updated settings, which
attempt to start the editor while it was closing.

* build: Update GutenbergKit version

* refactor: Address lint errors

* build: Update GutenbergKit version