CI: Add assemble android test jobs for both wordpress and jetpack
This is done because there has been now multiple occasions where a
dependency got updated, with the 'androidTest' source set failing to
compile, but since CI is no longer running instrumented tests, such
failure are not reported on the spot.

Previous Example: [Tests] Fix Android Tests Compilation Failure #21627

The main problem with this is that when this happens, at least our
weekly scheduled 'Dependency Analysis' job starts failing, and this
causes us too much trouble to get to the bottom of this. An
extra PR needs created/reviewed/tested, just to fix this small failure,
which would be otherwise fixed on the spot.