CI: Run assemble android test as part of and after a prototype build
Running the 'assembleAndroidTest' task as part of a separate job is too
slow and resource intensive. Running the 'assembleAndroidTest' task as
part of and after a "Prototype Build", which would alread have already
run the 'assemble' task before and thus have everything set might make
the whole process much faster.