fix: Prevent Media Library null pointer exceptions
Specific reproduction steps were elusive, but opening the media library
in the GutenbergKit editor sometimes led to a crash. Specifically, it
occurred after running the app on a newly-created emulator.

This change also updates the logic to align with guard logic found in
other portions of the codebase--e.g., PublishSettingsViewModel.