fix: Mitigate pending editor content loss (#22186)
* fix: Preserve GutenbergKit WebView state during config changes

Follow Google recommendations to disable restarting an activity that
house WebViews. Instead, handle the events in the activity itself. This
helps ensure state like content and scroll position persist through
these configuration changes.

https://developer.android.com/develop/ui/compose/quick-guides/content/manage-webview-state

* fix: Reinstate EditPostActivity's previous `configChanges` settings

This was modified for GutenbergKit's WebView, but is no longer necessary
now that GutenbergKit is managed by a separate GutenbergKitActivity.

https://github.com/wordpress-mobile/WordPress-Android/pull/21662