Remove check for modified main `strings.xml` outside release branch
This Danger check was out of place as
`WordPress/src/main/res/values/strings.xml` is the source of truth for
localizations.

Unlike in Apple projects, where we use `genstrings` to find localized
strings in the codebase to assemble the `strings` file, in Android, as
far as I understand, `<app>/src/main/res/valudes/strings.xml` is the
source of truth.

I believe this check ended up in `Dangerfile` accidentally. It was
introduced in
https://github.com/wordpress-mobile/WordPress-Android/pull/20132.
The update was done across our suite of apps, see for example
https://github.com/woocommerce/woocommerce-android/pull/10692 and
https://github.com/woocommerce/woocommerce-ios/pull/11926. Given I
couldn't find a comment explicitly about adopting this check, I guess
neither @iangmaia nor myself notice the same pattern as iOS was adopted
in the Android repo.

To reinforce the idea that this check is out of place, notice that
WooCommerce Android does not implement the check. _Of course, it could
be that WooCommerce is wrong, but lacking code to update `strings.xml`
as part of the code freeze process, that explanation is not satisfying._